extract App component from index.js (makes it easier to upgrade create-react-app)
